+++ lib/main.dart
@@ -1,332 +1,23 @@
 import 'package:flutter/material.dart';
-import 'package:image_picker/image_picker.dart';
-import 'dart:io';
+import 'screens/registration_page.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       title: 'Student Profile App',
       theme: ThemeData(
-        primarySwatch: Colors.blue,
+        primarySwatch: Colors.deepPurple,
+        visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
-      home: StudentRegistrationPage(),
-    );
-  }
-}
-
-class StudentRegistrationPage extends StatefulWidget {
-  @override
-  _StudentRegistrationPageState createState() =>
-      _StudentRegistrationPageState();
-}
-
-class _StudentRegistrationPageState extends State<StudentRegistrationPage> {
-  @override
-  final _formKey = GlobalKey<FormState>();
-
-  // Controllers for TextFields
-  final TextEditingController _nameController = TextEditingController();
-  final TextEditingController _emailController = TextEditingController();
-  final TextEditingController _contactController = TextEditingController();
-  final TextEditingController _dobController = TextEditingController();
-
-// Variables for Dropdown and Image
-  String? _selectedGender;
-  String? _profileImagePath;
-
-  File? _profileImage;
-  final ImagePicker _picker = ImagePicker();
-
-  Future<void> _pickImage() async {
-    final XFile? image = await _picker.pickImage(source: ImageSource.gallery);
-    if (image != null) {
-      setState(() {
-        _profileImage = File(image.path);
-      });
-    }
-  }
-
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.deepPurple, // Custom color
-        centerTitle: true, // Optional: center the title
-        title: Text(
-          'Student Profile App',
-          style: TextStyle(
-            fontSize: 24, // Make title bigger
-            fontWeight: FontWeight.bold, // Make title bold
-            color: Colors.white, // Set title color
-          ),
-        ),
-      ),
-      body: SingleChildScrollView(
-          padding: EdgeInsets.all(16),
-          child: Form(
-            key: _formKey,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  'Student Registration',
-                  style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-                ),
-                SizedBox(height: 16),
-
-                Center(
-                  child: Column(
-                    children: [
-                      CircleAvatar(
-                        radius: 50,
-                        backgroundImage: _profileImage != null
-                            ? FileImage(_profileImage!)
-                            : null,
-                        child: _profileImage == null
-                            ? Icon(Icons.person, size: 50)
-                            : null,
-                      ),
-                      TextButton.icon(
-                        onPressed: _pickImage,
-                        icon: Icon(Icons.photo),
-                        label: Text('Upload Profile Picture'),
-                      ),
-                    ],
-                  ),
-                ),
-                SizedBox(height: 16),
-
-                // Full Name Field
-                TextFormField(
-                  controller: _nameController,
-                  decoration: InputDecoration(
-                    labelText: 'Full Name',
-                    border: OutlineInputBorder(),
-                  ),
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return 'Please enter your full name';
-                    }
-                    return null;
-                  },
-                ),
-
-                SizedBox(height: 16),
-
-                // Email Field
-                TextFormField(
-                  controller: _emailController,
-                  decoration: InputDecoration(
-                    labelText: 'Email',
-                    border: OutlineInputBorder(),
-                  ),
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return 'Please enter your email';
-                    } else if (!RegExp(r'^[\w-.]+@([\w-]+\.)+[\w]{2,4}$')
-                        .hasMatch(value)) {
-                      return 'Enter a valid email';
-                    }
-                    return null;
-                  },
-                ),
-
-                SizedBox(height: 16),
-
-                // Contact Number Field
-                TextFormField(
-                  controller: _contactController,
-                  decoration: InputDecoration(
-                    labelText: 'Contact Number',
-                    border: OutlineInputBorder(),
-                  ),
-                  keyboardType: TextInputType.phone,
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return 'Please enter your contact number';
-                    }
-                    return null;
-                  },
-                ),
-                // Date of Birth Field
-                SizedBox(height: 16),
-                TextFormField(
-                  controller: _dobController,
-                  readOnly: true,
-                  decoration: InputDecoration(
-                    labelText: 'Date of Birth',
-                    border: OutlineInputBorder(),
-                    suffixIcon: Icon(Icons.calendar_today),
-                  ),
-                  onTap: () async {
-                    DateTime? pickedDate = await showDatePicker(
-                      context: context,
-                      initialDate: DateTime.now(),
-                      firstDate: DateTime(1950),
-                      lastDate: DateTime.now(),
-                    );
-                    if (pickedDate != null) {
-                      setState(() {
-                        _dobController.text =
-                            "${pickedDate.toLocal()}".split(' ')[0];
-                      });
-                    }
-                  },
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return 'Please select your date of birth';
-                    }
-                    return null;
-                  },
-                ),
-
-// Gender Dropdown
-                SizedBox(height: 16),
-                DropdownButtonFormField<String>(
-                  value: _selectedGender,
-                  decoration: InputDecoration(
-                    labelText: 'Gender',
-                    border: OutlineInputBorder(),
-                  ),
-                  items: ['Male', 'Female', 'Other'].map((String gender) {
-                    return DropdownMenuItem<String>(
-                      value: gender,
-                      child: Text(gender),
-                    );
-                  }).toList(),
-                  onChanged: (String? newValue) {
-                    setState(() {
-                      _selectedGender = newValue;
-                    });
-                  },
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return 'Please select your gender';
-                    }
-                    return null;
-                  },
-                ),
-
-                SizedBox(height: 24),
-
-                Center(
-                  child: ElevatedButton(
-                    onPressed: () {
-                      if (_formKey.currentState!.validate()) {
-                        if (_dobController.text.isEmpty) {
-                          ScaffoldMessenger.of(context).showSnackBar(
-                            SnackBar(
-                                content: Text('Please select date of birth')),
-                          );
-                        } else if (_selectedGender == null) {
-                          ScaffoldMessenger.of(context).showSnackBar(
-                            SnackBar(content: Text('Please select gender')),
-                          );
-                        } else if (_profileImage == null) {
-                          ScaffoldMessenger.of(context).showSnackBar(
-                            SnackBar(
-                                content:
-                                    Text('Please upload a profile picture')),
-                          );
-                        } else {
-                          // All fields valid: Navigate to display screen
-                          Navigator.push(
-                            context,
-                            MaterialPageRoute(
-                              builder: (context) => ProfileDisplayPage(
-                                name: _nameController.text,
-                                email: _emailController.text,
-                                gender: _selectedGender!,
-                                dob: _dobController.text,
-                                contact: _contactController.text,
-                                profileImage: _profileImage,
-                              ),
-                            ),
-                          );
-                        }
-                      }
-                    },
-                    child: Text('Submit'),
-                  ),
-                ),
-              ],
-            ),
-          )),
-    );
-  }
-}
-
-class ProfileDisplayPage extends StatelessWidget {
-  final String name;
-  final String email;
-  final String contact;
-  final String dob;
-  final String gender;
-  final File? profileImage;
-
-  const ProfileDisplayPage({
-    Key? key,
-    required this.name,
-    required this.email,
-    required this.gender,
-    required this.dob,
-    required this.contact,
-    this.profileImage,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(
-          'Profile Display',
-          style: TextStyle(fontSize: 22, fontWeight: FontWeight.bold),
-        ),
-        backgroundColor: Colors.deepPurple,
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(20.0),
-        child: Column(
-          children: [
-            Center(
-              child: CircleAvatar(
-                radius: 50,
-                backgroundImage:
-                    profileImage != null ? FileImage(profileImage!) : null,
-                child:
-                    profileImage == null ? Icon(Icons.person, size: 50) : null,
-              ),
-            ),
-            SizedBox(height: 20),
-            infoRow('Name:', name),
-            infoRow('Email:', email),
-            infoRow('Contact:', contact),
-            infoRow('Date of Birth:', dob),
-            infoRow('Gender:', gender),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget infoRow(String label, String value) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 8),
-      child: Row(
-        children: [
-          Text(
-            label,
-            style: TextStyle(fontWeight: FontWeight.bold),
-          ),
-          SizedBox(width: 10),
-          Flexible(child: Text(value)),
-        ],
-      ),
+      home: const StudentRegistrationPage(),
     );
   }
 }
